comit with database data

+++ backend/app/Http/Controllers/CompanyController.php
@@ -22,6 +22,16 @@ class CompanyController extends Controller
     {
         return response()->json(Company::all(),200);
     }
+    
+    public function getActiveCompanies(){
+        $companies = Company::where('active', '1') -> get();
+        return response()->json($companies,200);
+    }
+
+    public function getArchivedCompanies(){
+        $companies = Company::where('active', '0') -> get();
+        return response()->json($companies,200);
+    }
 
     public function getCompanyById($id)
     {
@@ -30,6 +40,13 @@ class CompanyController extends Controller
             return response()->json(['message' => 'Korisnik nije registrovan'], 404);
         }
         return response()->json($company, 200);
+    }
+
+    public static function getCompanyName($id){
+
+        $company = Company::findOrFail($id);
+        $companyName = $company['name'];
+        return $companyName;
     }
 
     /**
@@ -51,7 +68,7 @@ class CompanyController extends Controller
     public function store(Request $request)
     {
         $company = new Company;
-        $company->companyName = $request->companyName;
+        $company->name = $request->name;
         
         if($company->save()){
             $response['status'] = 1;
@@ -111,13 +128,70 @@ class CompanyController extends Controller
     public function destroy($id)
     {
         $company = Company::find($id);
-        $company->delete();
+        if(is_null($company)){
+
+            $response['status'] = 2;
+            $response['code'] = 404;
+            $response['message'] = 'Kompanija nije registrovana';
+
+            return response()->json($response);
+        }else{
+            $company->delete();
+
+            $response['status'] = 1;
+            $response['code'] = 200;
+            $response['message'] = 'Uspešno ste uklonili kompaniju';
+
+        return response()->json($response);
+        }
+        
     }
 
-    public function delete($id)
+    public function delete(Request $request, $id)
     {
         $company = Company::findOrFail($id);
-        $company->active = 0;
-        $company->update();
+
+        if(is_null($company)){
+
+            $response['status'] = 2;
+            $response['code'] = 404;
+            $response['message'] = 'Kompanija nije registrovana';
+
+            return response()->json($response);
+        }else{
+            $company->active = '0';
+            $company->save();
+
+            $response['status'] = 1;
+            $response['code'] = 200;
+            $response['message'] = 'Uspešno ste uklonili kompaniju';
+
+        return response()->json($response);
+        }
     }
+
+    public function activateCompany(Request $request, $id)
+    {
+        $company = Company::findOrFail($id);
+
+        if(is_null($company)){
+
+            $response['status'] = 2;
+            $response['code'] = 404;
+            $response['message'] = 'Kompanija nije registrovan';
+
+            return response()->json($response);
+        }else{
+            $company->active = '1';
+            $company->save();
+
+            $response['status'] = 1;
+            $response['code'] = 200;
+            $response['message'] = 'Uspešno aktivirali kompaniju';
+
+        return response()->json($response);
+        }
+    }
+
+    
 }
